Update common_js.js
using javascript to add elements when the game is finished to explain who won and provide a button for resetting the game.  these elements are removed by looping over the '.removable' class and removing the click event for playing a round with the designated choice.

ending of the game will also disable the click event from the Rock Paper Scissors representation. Resetting the game adds the click event back to the elements.